[ProfilesReordering] Create the profile order pref
Creating 'profile.profiles_order' which purpose is to store the profile
order to be displayed when showing all profiles (e.g. the profile
picker).
Currently the list is only updated when a profile is created/removed.
When creating a profile, by default we add the profile to the end of
the list (since we do not sort by profile name anymore).
In future CLs, there will be possible modifications within the list.

Bug: 1466146
Change-Id: I0fb80603a0ace984819cb331853dad44dd71e2f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4714858
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1176617}