[iOS] Fix sign-in in progress in SigninPromoMediator
This patch resolve a DCHECK signs in with syncpassphrase and then
turns on sync in bookmark.

The issue was sync service notifications was received before the
sign-in was done. Those notifications were setting
`self.signinInProgress` to NO.

To fix the issue, `signinInProgress` becomes private and is only set
when sign-in starts and stops.
Also private `initialSyncInProgress` property is introduced to track
the initial sync. It is set to YES (if needed) when sign-in is done.

For the public `signinInProgress` is replaced by
`showSigninOrInitialSyncSpinner`. This value is YES if either
`signinInProgress` or `initialSyncInProgress` is YES.

Fixed: 1467116
Change-Id: Ie7332d4538096fe5d610b44a22f561a464785b4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4718589
Reviewed-by: Menghan Yang <myuu@google.com>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176618}