[ios] Add GetOffTheRecordStatePath() method to ChromeBrowserState
The SessionRestorationService will manage session for both regular
and off-the-record ChromeBrowserState. It needs to know the state
path of the off-the-record ChromeBrowserState without creating it.

Move the implementation of GetStatePath() to ChromeBrowserState,
removing duplicated code in all sub-classes (and a copy of the
off-the-record ChromeBrowserState state path in the regular one).

Bug: 1383087
Change-Id: If152175ca9aece96bce96eb186e052bd8ef0207d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727385
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176637}