WebUI CR2023: Fix cr-toggle to work in RTL mode.
Essentially re-implemented the CSS logic for positioning the knob, so
that it is easier to calculate its position in RTL or LTR mode. Flipping
between the two is as simple as changing a CSS variable from 1 to -1.
The new logic relies more on 'translate()' and less on 'left' property.

Fixed: 1468198
Change-Id: I4db3d6bec31daa851d3998de9bd17c6b89b867c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4728463
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Auto-Submit: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176667}