AW: strictly require AwJUnit4ClassRunner
WebView tests require AwJUnit4ClassRunner in order to respect
"--disable-field-trial-config" (as well as some other requirements for
this test runner, such as running in process modes). However it was
possible to add a new test class without any @RunWith annotation.

This makes the presubmit check stricter. This also required renaming
DisableHardwareAccelerationForTest to be simply
DisableHardwareAcceleration, to make it clear this is a utility and not
actually a junit test.

Another impact of this change is that AwPageLoadMetricsTest will now run
in both process modes (single process and multiprocess), which is
consistent with all other WebView instrumentation tests.

Bug: 1468352
Test: run_webview_instrumentation_test_apk -f AwPageLoadMetricsTest.*
Test: git cl presubmit --upload (tested wrong @RunWith and w/o @RunWith)
Change-Id: I7dbf93b6a96b24caad13fa2954a95a829cadea2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727465
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176680}