[ntp-modules] V2 Disable & Dismiss affordances
Adds logic to support disabling and dismissing of modules in the V2
modules container.

Delegates the responsibility of insertion and removal of wrapper element
DOM nodes to the ModulesV2Element and leverages template for bindings.
This replaces the original use of the dom-repeat component, and allows
for a more efficient manipulation of the container's children nodes for
removal and/or insertion of elements as part of the dismiss affordance.

See design doc for additional details:
https://docs.google.com/document/d/1xhHGj1S9bVEMCBCMK8aTrv5_BzN_AFbUd1iKJSSYTCI/edit?resourcekey=0-OLgclhMYg2klzwKJw8k6-w#bookmark=id.rhuiyn5lg66n

Demo(s):
https://drive.google.com/file/d/1v35b3zHjSSc-GehyxHRymWpRu-Ub_Qs1/view

Bug: 1446716
Change-Id: Ibbcb1961120909df1b75ee821db70ba61936c853
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4572937
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Roman Arora <romanarora@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176681}