[iOS] Friendlier Safe Browsing Nits
We clean up code here to remove double flag guards and remove a
function that is only used once in the new code. Instead, code in the
mediator will be behind the related feature flag. Additionally, we
add enhanced protection as a visible flag on iOS devices.

Bug: 1457655
Change-Id: I7c84acc51730ae2e3332ae003c064a06183fdf44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727189
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176682}