Add color space support to NdkVideoEncodeAccelerator
This follows the model on other platforms where we recreate the
underlying encoder if the frame color space changes. It seems
to work well on newer devices, but not always on older ones.

In cases where it doesn't work, it seems the Android decoder
just assumes a color space instead of saying unspecified.

Bug: 1377842
Fixed: 1291779

Change-Id: I3ccae26dd96a09a90cb431abde005298866ae00a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4485120
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176686}