[scheduler] Pass MainThreadTaskQueue by reference to OnTaskCompleted()
Since graceful shutdown has been moved to the blink scheduler in
crrev.com/c/4349232, MainThreadTaskQueue is guaranteed to remain alive
while running tasks. As a follow-up, this CL changes OnTaskStarted(),
OnTaskCompleted(), and related methods to take a reference to MTTQ. This
also removes the "detached" queue type which was used in UKM as an
unknown type for detached queues.

Bug: 1143007
Change-Id: I849971d9e8bfdeadc757658a196bffa0dfb5de0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4605783
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176687}