[signin] Provide a specific signin error callback to the profile picker
This CL is a refactoring without behavior change. It will allow passing
a specific callback for the profile picker in a future CL.
It makes the architecture consistent with the EnableSyncCallback.

It also fixes a crash because the ProfilePicker tried to load the NTP,
which is not supported.

Bug: 1467483
Change-Id: Ia66b1872e47c36764cb88a7bd91fbd90165e184e
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4714890
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Auto-Submit: David Roger <droger@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176690}