Introduce TypedScopedUserManager and its Reset/Get.
On using ScopedUserManager, in many cases, we keep a typed raw pointer
as next to ScopedUserManager. Also, ScopedUserManager is not reset-table
it is used in std::unique_ptr<ScopedUserManager> form.

This CL allows user code to simply write
TypedScopedUserManager<FakeUserManager> etc., which reduces
some syntax-noise/boilarplates.

As an example, BrowserManagerTest starts to use it.

BUG=b:278643115
TEST=Tryjob

Change-Id: I4ac53d195a51061bb4dd8323c0c3f8e0af3ca247
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4728598
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176701}