[tab search] Themify favicons where necessary
Tab Search leans on favicons provided by TabRenderData for the
entries in its tab list. In many cases an image model is returned
that uses a ColorProvier to rasterize on demand, allowing the
favicon to be themed appropriately for a given theme state.

However many favicons are provided as pre-rasterized images directly
from the web content. In these situations the ColorProvider is not
available at the time the favicons are rasterized and the resulting
images are theme-unaware.

To solve the issue of these favicons looking out of place in a themed
Chrome browser, TabRenderData also reports whether a favicon's colors
should be adjusted to better contrast with browser UI. This is
Tracked in TabRenderData::should_themify_favicon.

This is set for many of Chrome's internal pages below
https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:chrome/browser/favicon/favicon_utils.cc;l=196;drc=714e1e34e823b7b4bc172fa4bdefee5aaf5e5fff

This CL applies the appropriate theme transformations where necessary
for Tab Search.

Bug: 1457578
Change-Id: I462d7fcbc546f59b6eed1c628947388a4b54ddc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4725308
Quick-Run: Alex Yang <aycyang@chromium.org>
Commit-Queue: Alex Yang <aycyang@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Alex Yang <aycyang@google.com>
Auto-Submit: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176722}