Add a missing run_loop.Run()
This has "accidentally worked" until now because
`url_loader_factory_.SimulateResponseForPendingRequest` calls
`RunUntilIdle`. However, with the switch to a new Rust-based JSON
parser, parsing happens in a thread pool and is not spotted by
`RunUntilIdle`. Adding the missing `run_loop.Run`, as it appears in the
other tests in this file, makes the test pass again.

Bug: 1455145
Change-Id: I2c615dcc79bd4d540f6d7201c255056941912ef3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727191
Commit-Queue: Dustin Mitchell <djmitche@chromium.org>
Reviewed-by: Russ Hamilton <behamilton@google.com>
Cr-Commit-Position: refs/heads/main@{#1176723}