[phrase-line-break] Implement the phrase line breaking
This patch implements the line breaking logic for the phrase
line breaking.

The logic uses the AdaBoost ML engine in the ICU.

If the line overflows, it falls back to normal line breaking.
If it still overflow, follow the normal CSS logic, as
specified by the `overflow-wrap` property. This behavior is a
feedback from the BudouX team. It's not clear whether this
behavior should be spec'ed or not, it's under discussion at:
https://github.com/w3c/csswg-drafts/issues/9033

Tests are not in wpt because the algorithm is UA-defined.

This patch has no behavior changes as all changes are under a
runtime flag.

Bug: 1443291
Change-Id: I0382bf42420dcadc9327824352d18615ae22c508
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4726301
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176755}