[iOS] Make WebClient responsible for enabling the Fullscreen API
Currently, ios/web decides for itself whether to enable the
Fullscreen API. However, properly supporting the API requires the
embedder to handle things like exiting fullscreen when it wants
to show a prompt.

This CL adds a WebClient method for enabling the Fullscreen API,
so that embedders can independently decide whether to enable it.

Bug: 1295480
Change-Id: Ibe331b28c5ff660ce1d0244fd15123b3aa9fbcba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4720678
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176772}