Use consumer deep scan UX when connector is report-only
For enterprises with the download connector enabled in report-only mode,
users can still start deep scans via the consumer prompt for scanning.
In that case we should still give them the consumer UX, even though the
connector is technically "enabled". This CL changes the string and
buttons for this situation.

Bug: 1424527
Change-Id: I828a691d72952b5b3ad9b2c240f4c6cec6f6db45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4728301
Reviewed-by: Lily Chen <chlily@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176775}