Disabling flaking CertVerifyProcInternalTest#TestKnownRoot
This appears to be due to a bug on the actual devices on the
android-cronet-x86-dbg-pie-tests bot, where the system clock is jumping
around, so we expect to re-enable this once this issue is fixed on the
bots' devices. But we can't continue to block the CQ with this issue,
so we are temporarily disabling.

Bug: 1468330
Change-Id: Ibf91350a03f83fc2009c3dec0f24bbc1a3d89646
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4729208
Auto-Submit: Sam Maier <smaier@chromium.org>
Commit-Queue: Sam Maier <smaier@chromium.org>
Commit-Queue: Matt Mueller <mattm@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176805}