Mac: don't show topchrome separator in immersive
Two issues here:
- The separator is redundant with the window shadow in light mode (see
screenshot in bug)
- The "thin view controller" which was added to work around
https://crbug.com/1369643 is stacked on top of the separator and very
visible. It's not clear to me if we still need this, and I'm working
on getting more details on how this was supposed to work. That said,
based on the description, it seems like the *size* of the controller
is the most important part, so making it transparent should be fine.

Bug: 1468415
Change-Id: Ie9c520de6c6c7bac6420a18ca568b3e6f93d13c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727010
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jayson Adams <shrike@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176821}