[Lacros A11y] Fix STS and Switch Access content script injection in Lacros
Google Docs was refusing to allow the EmbeddedA11yHelper extension
to inject the <script> tag with an error that "This document requires
'TrustedHTML' assignment". This only happened in Lacros, which uses
MV3 to inject the script, and not Ash, which uses MV2. Updates the
injected script to work properly in Lacros by using a TrustedScript.

Note that using "Eval" in the Ash extensions does not seem to work,
so we continue to add a <script> tag in Ash.

Test: Manual in Lacros, Ash on Chrome OS on Linux
Bug: b/293466039
Change-Id: Ic3c91926e3aab36a4718dd68ea0a5df01ef832a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4730018
Commit-Queue: Josiah Krutz <josiahk@google.com>
Reviewed-by: Josiah Krutz <josiahk@google.com>
Auto-Submit: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176830}