[Extensions] Work around dangling BrowserContext pointer.
WebRequestAPI binds a callback to some ExtensionWebRequestEventRouter
member functions and one of those parameters is a BrowserContext. It
then posts that callback to a SequenceTaskRunner.

However, it's possible for that BrowserContext to be destroyed before
the task runs, which would result in a UAF error when the caller uses
that BrowserContext. The fix is for the WebRequestAPI instance to
bind to new member functions that call through to the target
functions. This callback is then bound using a WeakPtr to prevent
the call if the BrowserContext is torn down. Additionally, the
forwarding functions check the validity of the BrowserContext before
forwarding the call.

This works because WebRequestAPI is a keyed service bound to the
lifetime of the BrowserContext and is destroyed before the
BrowserContext is.

Bug: 1453528, 1468177
Change-Id: I7667cd9ec0247d29b45aa17ec965201f4bb4f064
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4726946
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Auto-Submit: David Bertoni <dbertoni@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176842}