media/gpu/v4l2: Handle End-of-Stream during mid stream
ToT code doesn't support EOS midstream, only at the end. Such
operations are not very common in the wild, but are a necessary part
of midstream resolution changes (which are very common). To simplify
code review, this CL focuses on the first part, and adds a specific
test case for it in v_d_a_test.

Specifically, to support these EOS midstream:
- We add a SendStartCommand() to V4L2Queue, and use it.
- Two new methods are added to V4L2StatefulVD to reuse operations,
namely DrainOUTPUTQueue() and PrintOutQueueStatesForVLOG().
 DrainOUTPUTQueue() is called during Decode() and also during the
final part of the Flush operation.

Tested via:

/usr/local/libexec/chrome-binary-tests/video_decode_accelerator_tests
/usr/local/share/tast/data_pushed/go.chromium.org/tast-tests/cros/local/bundles/cros/video/data/test-25fps.vp8
--vmodule=v4l2_stateful_video_decoder=3,*/media/gpu/v4l2/*=1,=0,video_decoder_pipeline=3
--enable-features=V4L2FlatStatefulVideoDecoder
--gtest_filter='*.FlushMidStream'

on hana and trogdor which are the only V4L2 stateful ChromeOS devices.

[1] https://www.kernel.org/doc/html/v5.15/userspace-api/media/v4l/dev-decoder.html#drain

Bug: b:279980150
Change-Id: I9a1c39e5a95fdd005f0aeb9051481a372a5ec726
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727841
Reviewed-by: Hang Nguyen <hnt@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176850}