[Extensions] Allow the user to remove multiple extensions at once
The API is used by the extension safety check when clicking on the
remove all button to show an uninstall dialog and uninstall those
extensions. If an extension fails to be uninstalled, it will not
pause the uninstall of the other extensions on the list. The UI
behavior of this multiple extension install dialog is similar to
the single extension dialog.

Screenshot for the UI change: https://screenshot.googleplex.com/3vbodFqexcXtMuu

Bug: 1432194
Change-Id: Ic5e3aedbc1cd851b3776f980afed228b3d403b54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4581775
Commit-Queue: Zack Han <zackhan@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176852}