Ensure content analysis completes before printing document
When the PrintingContext starts printing a new document, it can cause a
file save dialog to be displayed to the user.  It is important that this
does not happen while content analysis is still in progress, since the
printing operation might not be allowed.

Update the content analysis state change logic to ensure that there is
not a print job that has already started.  While existing tests pass
this check, a new test to cover window.print() will need work to satisfy
this since that use case is generating a file save dialog for users.

Bug: 1464566
Change-Id: I7d478a442e313a1fefd422649efa8b202053e8cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4704792
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176858}