bento_button: Fix crash for desk switch button tooltip
The crash occurs when the desk switch button tooltip appears during the
desk switch animation. The root cause is that we update the switch
buttons after the desk switch, which means the switch button may remain
visible for longer than expected.

To ensure that the switch buttons are valid throughout their lifetime,
the solution is to update them before the desk switch.

Bug: b/292257858
Test: Manual
Change-Id: I5a82e535a228f3f5a1390aeefb5d410dc7c660a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4718311
Reviewed-by: Daniel Andersson <dandersson@chromium.org>
Commit-Queue: Yongshun Liu <yongshun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176874}