gpu: win: Allow VisualSubtree properties to depend on multiple inputs
This CL splits the diffing of VisualSubtree::Update inputs and the
processing of the DComp visuals that result from it. This lets us have
multiple visual property updates that depend on the same inputs in a
simple way. Previously |content_soft_borders_| was used to track
changes to the content visual's border mode, but it is no longer needed
after this change.

Change-Id: Ia534d76464b822a91282ea944fe924f1db750201
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4720149
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Commit-Queue: Michael Tang <tangm@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1176891}