ash: fix potential uaf in DemoSessionMetricsRecorder
This code uses base::ScopedObservation to observe a Window. It
may call Observe() multiple times, but it does not necessarily call
Reset() before calling Observe(). ScopedObservation assumes
Reset() is called before Observe() (it has a DCHECK). If Reset() is
not called before Observe(), then the existing observer is not removed
before adding the new one.

This patch adds a call to Reset() to ensure it is not already observing
a window.

Bug: 1462483
Change-Id: Ie4ad0989a24e41381c87e2f7b51aa0ad2ef6c735
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727945
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176940}