Replace assets for launcher sort nudge and continue section notice
The new illustrations contain several color tokens (and the colors in
the illustrations dynamic based on the current theme). This requires
changing the images in the launcher toast views to be changed from
vector icons to lottie files.

The illustrations use several new color tokens, which needed to be added
to lottie color mappings.

BUG=b:288486159,b:289133234

Change-Id: Iec164e41eb57c79182fddb6ab9cb4149c314a214
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4722906
Reviewed-by: Sean Kau <skau@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176978}