cros_next: fixed shelf unittest failure
This cl fixed the failure of shelf manager unit test due to the
crash of the overview item shadow.

The crash is due to the small test window which is 1x370 which will
not appear in real case. The small size leads to a negative adjusted
shadow elevation and finally make the shadow layer smaller than its
nine patch layer's border.

The cl fixed the issue by setting a more meaningful size to test
window.

Test: ShelfLayoutManagerTest.VisibleInOverview
Bug: b:293400777
Change-Id: I479a562dfe71e241982b7dbf43b8c10ea8f5bd86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727783
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1177062}