Fix BrowsingDataModelTest with CompressionDictionaryTransportBackend
Currently BrowsingDataModelTest.ConcurrentDeletions and
BrowsingDataModelTest.DelegateDataDeleted fails when
CompressionDictionaryTransportBackend feature is enabled. This is
because NetworkContext::GetSharedDictionaryUsageInfo() IPC is called
if the feature is enabled.

To fix this failure, this CL adds EXPECT_CALL() of
GetSharedDictionaryUsageInfo() method.

The behavior of the feature itself is tested by
BrowsingDataModelSharedDictionaryTest in the test file below.

Bug: 1413922

Change-Id: Ic0b8664fdfb57eaad4b4334312eb733eb8356c88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4731152
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1177191}