[iOS] Restrict the Fullscreen API to tablet-only
Experimental data suggests that enabling the Fullscreen API on iPhone
leads to a worse user experience, likely because the iOS web
ecosystem assumes that this API is only available on iPad, since
Safari only ships this on iPad. This CL restricts this API to iPad.

Bug: 1295480
Change-Id: I96a78de82f3fe710ed2011a33dca1f4770ff0623
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727200
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1177282}