Cookie Settings: Do not add lifetime to incognito exceptions
When creating a User Bypass exception, do not set an expiration if the
associated profile is incognito. Incognito exceptions are cleared at the
end of the session anyway.

Making this change within CookieSettings means that the various UI
surfaces which display the expiration now display the exception as
permanent.

While this is not as good as a dedicated "This only lasts in incognito"
string, it's much better than the current "Expires in X days".

Bug: 1446230
Change-Id: Ie6e2ef8e28ca98824b0481f6756e731943ad61d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4732991
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1177302}